+++ react-container/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.6.0 Jan 22 2020
+
+### Changes
+
+- Make logs a bit safer and better
+- Use new authResult from `auth-utils`
+- Use new way of setup for CogniteAuth
+
 ## 0.4.0 Dec 23 2020
 
 ### Changes
@@ -23,4 +31,3 @@
 
 - exposed `retrieveAccessToken` so you can still get token directly
 - added `isLocalhost`
--